requirements_linux.txt
numpy
pillow
pyheif
scikit-learn
scipy
deepface